+++ packages/@expo/metro-config/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### 💡 Others
 
+## 54.0.6 — 2025-10-01
+
+### 💡 Others
+
 - Rename to `@expo/server` to `expo-server` ([#40087](https://github.com/expo/expo/pull/40087) by [@kitten](https://github.com/kitten))
 
 ## 54.0.5 — 2025-09-22